Add casts to avoid compiler warnings

+++ src/git2r_blob.c
@@ -1,6 +1,6 @@
 /*
  *  git2r, R bindings to the libgit2 library.
- *  Copyright (C) 2013-2024 The git2r contributors
+ *  Copyright (C) 2013-2026 The git2r contributors
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License, version 2,
@@ -339,5 +339,5 @@ cleanup:
     if (error)
         git2r_error(__func__, git_error_last(), NULL, NULL);
 
-    return Rf_ScalarInteger(size);
+    return Rf_ScalarInteger((int)size);
 }
